restructured code & type improve
