+++ src/clients/rpc-client/kyve/registry.v1beta1.ts
@@ -13,18 +13,27 @@ import {
 } from "@kyve/proto/dist/proto/kyve/registry/v1beta1/tx";
 import { coins } from "@cosmjs/stargate";
 import { StdFee } from "@cosmjs/amino/build/signdoc";
-import { withTypeUrl } from "../registry/tx.registry";
+import { withTypeUrl } from "../../../registry/tx.registry";
 import { AccountData } from "@cosmjs/amino/build/signer";
 import { BigNumber } from "bignumber.js";
-import { KYVE_DECIMALS } from "../constants";
-import { MessageEvent } from "../types/events";
+import { KYVE_DECIMALS } from "../../../constants";
+import { MessageEvent } from "../../../types/events";
 import { Tendermint34Client } from "@cosmjs/tendermint-rpc";
 import { toHex } from "@cosmjs/encoding";
 import { sha256 } from "@cosmjs/crypto";
-import { FullDecodedTransaction } from "../types/transactions";
+import { FullDecodedTransaction } from "../../../types/transactions";
 import { decodeTxRaw } from "@cosmjs/proto-signing";
-import { Client } from "../types/client";
-import { DENOM } from "../constants";
+import { Client } from "../../../types/client";
+import { DENOM } from "../../../constants";
+import { TextProposal } from "@kyve/proto/dist/proto/cosmos/gov/v1beta1/gov";
+import { ParameterChangeProposal } from "@kyve/proto/dist/proto/cosmos/params/v1beta1/params";
+import {
+  CancelPoolUpgradeProposal,
+  PausePoolProposal,
+  SchedulePoolUpgradeProposal,
+  UnpausePoolProposal,
+  UpdatePoolProposal,
+} from "@kyve/proto/dist/proto/kyve/registry/v1beta1/gov";
 
 export default class KyveBaseMsg {
   private nativeClient: Client;
@@ -443,10 +452,157 @@ export default class KyveBaseMsg {
 
     return events;
   }
+
+  private createGovTx(
+    amount: string,
+    content: { typeUrl: string; value: Object }
+  ) {
+    return {
+      typeUrl: "/cosmos.gov.v1beta1.MsgSubmitProposal",
+      value: {
+        content,
+        initialDeposit: coins(amount.toString(), "tkyve"),
+        proposer: this.account.address,
+      },
+    };
+  }
+
+  public submitTextProposal(
+    amount: string,
+    value: TextProposal,
+    fee?: StdFee | "auto" | number,
+    memo?: string
+  ) {
+    const content = {
+      typeUrl: "/cosmos.gov.v1beta1.TextProposal",
+      value: TextProposal.encode(value).finish(),
+    };
+    const tx = this.createGovTx(amount, content);
+    return this.nativeClient.signAndBroadcast(
+      this.account.address,
+      [tx],
+      fee ? fee : "auto",
+      memo
+    );
+  }
+
+  public parameterChangeProposal(
+    amount: string,
+    value: ParameterChangeProposal,
+    fee?: StdFee | "auto" | number,
+    memo?: string
+  ) {
+    const content = {
+      typeUrl: "/cosmos.params.v1beta1.ParameterChangeProposal",
+      value: ParameterChangeProposal.encode(value).finish(),
+    };
+    const tx = this.createGovTx(amount, content);
+    return this.nativeClient.signAndBroadcast(
+      this.account.address,
+      [tx],
+      fee ? fee : "auto",
+      memo
+    );
+  }
+
+  public updatePoolProposal(
+    amount: string,
+    value: UpdatePoolProposal,
+    fee?: StdFee | "auto" | number,
+    memo?: string
+  ) {
+    const content = {
+      typeUrl: "/kyve.registry.v1beta1.UpdatePoolProposal",
+      value: UpdatePoolProposal.encode(value).finish(),
+    };
+    const tx = this.createGovTx(amount, content);
+    return this.nativeClient.signAndBroadcast(
+      this.account.address,
+      [tx],
+      fee ? fee : "auto",
+      memo
+    );
+  }
+
+  public pausePoolProposal(
+    amount: string,
+    value: PausePoolProposal,
+    fee?: StdFee | "auto" | number,
+    memo?: string
+  ) {
+    const content = {
+      typeUrl: "/kyve.registry.v1beta1.PausePoolProposal",
+      value: PausePoolProposal.encode(value).finish(),
+    };
+    const tx = this.createGovTx(amount, content);
+    return this.nativeClient.signAndBroadcast(
+      this.account.address,
+      [tx],
+      fee ? fee : "auto",
+      memo
+    );
+  }
+
+  public unpausePoolProposal(
+    amount: string,
+    value: UnpausePoolProposal,
+    fee?: StdFee | "auto" | number,
+    memo?: string
+  ) {
+    const content = {
+      typeUrl: "/kyve.registry.v1beta1.UnpausePoolProposal",
+      value: UnpausePoolProposal.encode(value).finish(),
+    };
+    const tx = this.createGovTx(amount, content);
+    return this.nativeClient.signAndBroadcast(
+      this.account.address,
+      [tx],
+      fee ? fee : "auto",
+      memo
+    );
+  }
+
+  public schedulePoolUpgradeProposal(
+    amount: string,
+    value: SchedulePoolUpgradeProposal,
+    fee?: StdFee | "auto" | number,
+    memo?: string
+  ) {
+    const content = {
+      typeUrl: "/kyve.registry.v1beta1.SchedulePoolUpgradeProposal",
+      value: SchedulePoolUpgradeProposal.encode(value).finish(),
+    };
+    const tx = this.createGovTx(amount, content);
+    return this.nativeClient.signAndBroadcast(
+      this.account.address,
+      [tx],
+      fee ? fee : "auto",
+      memo
+    );
+  }
+
+  public cancelPoolUpgradeProposal(
+    amount: string,
+    value: CancelPoolUpgradeProposal,
+    fee?: StdFee | "auto" | number,
+    memo?: string
+  ) {
+    const content = {
+      typeUrl: "/kyve.registry.v1beta1.CancelPoolUpgradeProposal",
+      value: CancelPoolUpgradeProposal.encode(value).finish(),
+    };
+    const tx = this.createGovTx(amount, content);
+    return this.nativeClient.signAndBroadcast(
+      this.account.address,
+      [tx],
+      fee ? fee : "auto",
+      memo
+    );
+  }
   async getKyveBalance() {
     const data = await this.nativeClient.getBalance(
       this.account.address,
-      DENOM
+      "tkyve"
     );
     return data.amount;
   }
